Add chars 0x00 - 0x1F to the font
- Some levels used chars < 0x20 as non-collapsible spaces, those would
  now show up as [?]
- I recently found out how making characters < 0x20 6 pixels wide
  doesn't work if they're missing from the font altogether

For this change, I had to make some changes to the scripts and the
git.txt format. git.txt worked by including the character itself on
a dedicated line, in literal form. Which causes a problem with the
line feed character (\n) because it's indistinguishable from all the
other line feeds in the file... The NUL character might also be
annoying to work with because some text editors immediately consider
the file invalid UTF-8.

Instead, characters MAY now also be represented in hexadecimal form,
starting with 0x, in the git.txt.